// src/enums/permission.enum.ts
export enum PermissionEnum {
    ALL = 'all',

    VIEW_CAR='view_car',
    CREATE_CAR = 'create_car',
    EDIT_CAR = 'edit_car',
    DELETE_CAR = 'delete_car',
    VERIFY_CAR = 'verify_car',

    VIEW_USERS = 'view_users',
    BAN_USER = 'ban_user',
    UNBAN_USER = 'unban_user',
    DELETE_USER ='delete_user',
    CREATE_MANAGER ='create_manager',
    MANAGE_SALON = 'manage_salon',

}